requirements.txt
beautifulsoup4==4.14.3
datasets==4.6.1
requests==2.32.5
pandas==3.0.1
numpy==2.4.2
torch==2.10.0
transformers==5.3.0
accelerate==1.13.0
huggingface_hub==1.6.0
safetensors==0.7.0
sentencepiece==0.2.1
tokenizers==0.22.2
gguf==0.18.0
mistral-common
python-dotenv
